move character creation page to pages folfer
