+++ src/components/loadCharacter/loadCharacter.spec.jsx
@@ -142,7 +142,7 @@ describe('Load Character Intergration', () => {
     wrapper.find('IconButton').simulate('click');
     wrapper.find('button').find('span[children="Mr Load"]').simulate('click');
 
-    expect(wrapper.find('LoadedCharacter').exists()).toBe(true);
+    expect(wrapper.find('UsePage').exists()).toBe(true);
     expect(wrapper.text()).toContain('Mr Load');
   });
 });
